requirements.txt
Django==3.1.3
django-csvimport==2.16
django-import-export==2.5.0


